fix(speaker-stats): rendering issues
