+++ react/features/speaker-stats/components/web/SpeakerStatsItem.js
@@ -1,37 +1,11 @@
 /* @flow */
 
-import { makeStyles } from '@material-ui/core/styles';
 import React from 'react';
 
 import { Avatar } from '../../../base/avatar';
 import { FACIAL_EXPRESSIONS } from '../../../facial-recognition/constants.js';
 
 import TimeElapsed from './TimeElapsed';
-
-const useStyles = makeStyles(theme => {
-    return {
-        item: {
-            height: 48
-        },
-        avatar: {
-            height: 32
-        },
-        expressions: {
-            paddingLeft: 29
-        },
-        placeholderColor: {
-            color: theme.palette.text03
-        },
-        time: {
-            padding: '2px 4px',
-            borderRadius: '4px'
-        },
-        dominant: {
-            backgroundColor: theme.palette.success02
-        }
-
-    };
-});
 
 /**
  * The type of the React {@code Component} props of {@link SpeakerStatsItem}.
@@ -80,10 +54,15 @@ type Props = {
     isDominantSpeaker: boolean,
 
     /**
+     * Styles for the item.
+     */
+    styles: Object,
+
+    /**
      * Invoked to obtain translated strings.
      */
     t: Function
-};
+}
 
 const SpeakerStatsItem = (props: Props) => {
     /**
@@ -92,15 +71,13 @@ const SpeakerStatsItem = (props: Props) => {
      * @inheritdoc
      * @returns {ReactElement}
      */
-
-    const classes = useStyles();
-    const hasLeftClass = props.hasLeft ? classes.placeholderColor : '';
-    const rowDisplayClass = `row ${hasLeftClass} ${classes.item}`;
+    const hasLeftClass = props.hasLeft ? props.styles.placeholderColor : '';
+    const rowDisplayClass = `row ${hasLeftClass} ${props.styles.item}`;
     const expressionClass = 'expression';
     const nameTimeClass = `name-time${
         props.showFacialExpressions ? ' name-time_expressions-on' : ''
     }`;
-    const timeClass = `${classes.time} ${props.isDominantSpeaker ? classes.dominant : ''}`;
+    const timeClass = `${props.styles.time} ${props.isDominantSpeaker ? props.styles.dominant : ''}`;
 
 
     const FacialExpressions = () => FACIAL_EXPRESSIONS.map(
@@ -109,7 +86,7 @@ const SpeakerStatsItem = (props: Props) => {
                     aria-label = { props.t(`speakerStats.${expression}`) }
                     className = {
                         `${expressionClass} ${
-                            props.facialExpressions[expression] === 0 ? classes.placeholderColor : ''
+                            props.facialExpressions[expression] === 0 ? props.styles.placeholderColor : ''
                         }`
                     }
                     key = { expression }>
@@ -118,15 +95,11 @@ const SpeakerStatsItem = (props: Props) => {
             )
     );
 
-    if (props.hidden) {
-        return <div key = { props.participantId } />;
-    }
-
     return (
         <div
             className = { rowDisplayClass }
             key = { props.participantId } >
-            <div className = { `avatar ${classes.avatar}` }>
+            <div className = { `avatar ${props.styles.avatar}` }>
                 <Avatar
                     className = 'userAvatar'
                     participantId = { props.participantId } />
@@ -146,7 +119,7 @@ const SpeakerStatsItem = (props: Props) => {
             </div>
             { props.showFacialExpressions
             && (
-                <div className = { `expressions ${classes.expressions}` }>
+                <div className = { `expressions ${props.styles.expressions}` }>
                     <FacialExpressions />
                 </div>
             )
