+++ react/features/speaker-stats/components/web/SpeakerStatsList.js
@@ -7,10 +7,29 @@ import abstractSpeakerStatsList from '../AbstractSpeakerStatsList';
 
 import SpeakerStatsItem from './SpeakerStatsItem';
 
-const useStyles = makeStyles(() => {
+const useStyles = makeStyles(theme => {
     return {
         list: {
             marginTop: 15
+        },
+        item: {
+            height: 48
+        },
+        avatar: {
+            height: 32
+        },
+        expressions: {
+            paddingLeft: 29
+        },
+        placeholderColor: {
+            color: theme.palette.text03
+        },
+        time: {
+            padding: '2px 4px',
+            borderRadius: '4px'
+        },
+        dominant: {
+            backgroundColor: theme.palette.success02
         }
     };
 });
@@ -21,8 +40,8 @@ const useStyles = makeStyles(() => {
  * @returns {React$Element<any>}
  */
 const SpeakerStatsList = () => {
-    const items = abstractSpeakerStatsList(SpeakerStatsItem);
     const classes = useStyles();
+    const items = abstractSpeakerStatsList(SpeakerStatsItem, classes);
 
     return (
         <div className = { classes.list }>
